Move example_site to separate package (avoid colliding with maigret.sites module) and update test import
